pyproject.toml
[tool.poetry]
name = "geneticfitness"
version = "0.1.0"
description = ""
authors = ["Tobia Ochsner <tobia.ochsner@hotmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.isort]
profile = "black"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.1.1"
numpy = "^1.26.0"
tqdm = "^4.66.1"
matplotlib = "^3.8.0"
seaborn = "^0.12.2"
scikit-learn = "^1.3.1"
black = "^23.9.1"
pyright = "^1.1.327"
isort = "^5.12.0"
flake8 = "^6.1.0"
pre-commit = "^3.4.0"
flake8-tidy-imports = "^4.10.0"
flake8-simplify = "^0.20.0"
biopython = "^1.81"
loguru = "^0.7.2"
viennarna = {path = "/Users/tobiaochsner/ViennaRNA-2.6.3"}

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
jupyter-black = "^0.3.4"
jupyterlab = "^4.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"